update analytics chart skeletons
