+++ src/components/skeletons/index.ts
@@ -2,3 +2,4 @@ export { default as Table } from "./table"
 export { default as EventLogs } from "./event-logs"
 export { default as RequestLogs } from "./request-logs"
 export { default as PolicyFields } from "./policy-fields"
+export * as AnalyticsCharts from "./analytics-charts"
